Issue #141 - team 5
A simple design for list of online users. I also added the toggle
animation for hidding it and the title for this 'box' (Active Users or
Logged in Users doesn't really sound good, but that's the least of my
worries :)).

Also added the ability to display only users that contain the searched
string and to arrange list of logged in users by alphabet(design needs
to be improved). Also the true image of the user should be passed in
'data' to javascript. We (I and igypigi->did most of the javascript)
were also unable to test this completely, because the updating logged
in users doesn't always work(on our computers) and we weren't able to
figure out why.